# YJIT: Generate specialized code for Symbol for objtostring