For YJIT: This adds specialized code for objtostring when dealing with Symbol. We're looking to get the same memory optimization that is done in the interpreter, [see this commit a while back in the interpreter](https://github.com/ruby/ruby/commit/b08dacfea39ad8da3f1fd7fdd0e4538cc892ec44).

The premise for the optimization is that during string interpolation (which doesn't require a mutable string), we can avoid extra allocations for Symbol by calling `rb_sym2str` to get its frozen string literal without any object allocations.

## Testing & Metrics
We [used this test to record the number of allocations](https://gist.github.com/jhawthorn/0c3844867e267949db39f90032827122) for to_s.

Results:
`./miniruby --yjit --yjit-call-threshold=1 test.rb`
Before: 3.4.0 master (af5c34fb0b6d06f9c78952bccc9879fa735b3857)
Output: 2000001
After:
Output: 1000001

## Performance
Rough performance testing with `time`, we can see improvements in total time:
**Before: 3.4.0 master (af5c34fb0b6d06f9c78952bccc9879fa735b3857) (YJIT disabled)**
`./miniruby test.rb  1.39s user 0.03s system 88% cpu 1.607 total`
**Before: 3.4.0 master (af5c34fb0b6d06f9c78952bccc9879fa735b3857) (YJIT enabled)**
`./miniruby --yjit --yjit-call-threshold=1 test.rb  1.42s user 0.03s system 87% cpu 1.646 total`

**After:**
`./miniruby --yjit --yjit-call-threshold=1 test.rb  0.89s user 0.02s system 94% cpu 0.957 total`

Co-authored with @jhawthorn